lesson-3-microservices/exercises/2_simultaneous_containers/starter/admin/requirements.txt
flask==2.2.2
psycopg2-binary
flask_sqlalchemy
